Refactor datetime_to_str to utils package
